refactor(vault): consolidate Member interface definitions
Refs: #170

- Add canonical Member interface to types.ts (single source of truth)
- Add MemberAuthContext = Pick<Member, ...> for permission checks
- Update permissions.ts to import from types.ts
  - Re-export Role and MemberAuthContext
  - Keep Member as alias for backward compatibility
- Update members.ts to import Member from types.ts

All 877 vault tests pass. No behavior changes.
